Move Session to its own file.
[ci skip-build-wheels]
