+++ src/rust/engine/src/lib.rs
@@ -40,12 +40,14 @@ mod intrinsics;
 mod nodes;
 mod scheduler;
 mod selectors;
+mod session;
 mod tasks;
 mod types;
 
 pub use crate::context::{Core, ExecutionStrategyOptions, RemotingOptions};
 pub use crate::core::{Failure, Function, Key, Params, TypeId, Value};
 pub use crate::intrinsics::Intrinsics;
-pub use crate::scheduler::{ExecutionRequest, ExecutionTermination, Scheduler, Session};
+pub use crate::scheduler::{ExecutionRequest, ExecutionTermination, Scheduler};
+pub use crate::session::Session;
 pub use crate::tasks::{Rule, Tasks};
 pub use crate::types::Types;
